Feature - Command Grouping
* Grouped commands by feature
* Cleaned up cog directory
